+++ aasrp/admin.py
@@ -3,10 +3,10 @@ Django admin declarations
 """
 
 # Django
-from django.contrib import admin
+from django.contrib import admin, messages
 
 # AA SRP
-from aasrp.models import AaSrpLink, AaSrpRequest, AaSrpRequestComment
+from aasrp.models import AaSrpLink, AaSrpRequest, AaSrpRequestComment, FleetType
 
 
 def custom_filter(title):
@@ -20,6 +20,26 @@ def custom_filter(title):
         """
         Custom_filter :: wrapper
         """
+
+        def expected_parameters(self):
+            """
+            Expected parameters
+            :return:
+            :rtype:
+            """
+
+            pass
+
+        def choices(self, changelist):
+            """
+            Choices
+            :param changelist:
+            :type changelist:
+            :return:
+            :rtype:
+            """
+
+            pass
 
         def __new__(cls, *args, **kwargs):
             instance = admin.FieldListFilter.create(*args, **kwargs)
@@ -108,3 +128,105 @@ class AaSrpRequestCommentAdmin(admin.ModelAdmin):
     list_display = ("srp_request", "comment_type", "creator")
     ordering = ("srp_request",)
     list_filter = ("comment_type",)
+
+
+@admin.register(FleetType)
+class FleetTypeAdmin(admin.ModelAdmin):
+    """
+    Config for fleet type model
+    """
+
+    list_display = ("id", "_name", "_is_enabled")
+    list_filter = ("is_enabled",)
+    ordering = ("name",)
+
+    @admin.display(description="Fleet Type", ordering="name")
+    def _name(self, obj):
+        """
+        Rewrite name
+        :param obj:
+        :type obj:
+        :return:
+        :rtype:
+        """
+
+        return obj.name
+
+    @admin.display(description="Is Enabled", boolean=True, ordering="is_enabled")
+    def _is_enabled(self, obj):
+        """
+        Rewrite is_enabled
+        :param obj:
+        :type obj:
+        :return:
+        :rtype:
+        """
+
+        return obj.is_enabled
+
+    actions = (
+        "activate",
+        "deactivate",
+    )
+
+    @admin.action(description="Activate selected fleet type(s)")
+    def activate(self, request, queryset):
+        """
+        Mark fleet type as active
+        :param request:
+        :type request:
+        :param queryset:
+        :type queryset:
+        :return:
+        :rtype:
+        """
+
+        notifications_count = 0
+        failed = 0
+
+        for obj in queryset:
+            try:
+                obj.is_enabled = True
+                obj.save()
+
+                notifications_count += 1
+            except Exception:
+                failed += 1
+
+        if failed:
+            messages.error(request, f"Failed to activate {failed} fleet types")
+
+        if queryset.count() - failed > 0:
+            messages.success(request, f"Activated {notifications_count} fleet type(s)")
+
+    @admin.action(description="Deactivate selected fleet type(s)")
+    def deactivate(self, request, queryset):
+        """
+        Mark fleet type as inactive
+        :param request:
+        :type request:
+        :param queryset:
+        :type queryset:
+        :return:
+        :rtype:
+        """
+
+        notifications_count = 0
+        failed = 0
+
+        for obj in queryset:
+            try:
+                obj.is_enabled = False
+                obj.save()
+
+                notifications_count += 1
+            except Exception:
+                failed += 1
+
+        if failed:
+            messages.error(request, f"Failed to deactivate {failed} fleet types")
+
+        if queryset.count() - failed > 0:
+            messages.success(
+                request, f"Deactivated {notifications_count} fleet type(s)"
+            )