+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning]
 ## [In Development] - Unreleased
 
 
+## [1.17.0] - 2023-03-06
+
+### Added
+
+- Fleet type to SRP links (optional) ([#83])
+- SRP details to SRP request form ([#84])
+
+[#83]: https://github.com/ppfeufer/aa-srp/issues/83 "[Feature Request] Add Fleet Types to SRP Links"
+[#84]: https://github.com/ppfeufer/aa-srp/issues/84 "[Feature Request] Add ARP Link Details to Request Form"
+
+
 ## [1.16.4] - 2022-11-26
 
 ### Fixed